added comments and some BONUS features

+++ server.js
@@ -1,13 +1,16 @@
+// Calling upon the dependencies used
 const cTable = require('console.table');
 const inquirer = require('inquirer');
 const mysql = require('mysql2');
+// To hide my password
 require('dotenv').config();
 
+// The ASCII Logo at the top created from info on the package.json
 const logo = require('asciiart-logo');
 const config = require('./package.json');
 console.log(logo(config).render());
 
-
+// Connecting to the SQL Database
 const connection = mysql.createConnection(
     {
         host: 'localhost',
@@ -18,7 +21,8 @@ const connection = mysql.createConnection(
     console.log(`Connected to employee_tracker_db`)
 );
 
-const questions = [{
+// Questions array that doesn't necessarily need to be an array but makes it easier to add questions if needed
+const question = [{
     type: "list",
     name: "what",
     message: "What would you like to do?",
@@ -31,12 +35,15 @@ const questions = [{
         'Add An Employee', 
         'Update An Employee Role', 
         'Update An Employee Manager',
+        'Delete A Department',
+        'Delete A Role',
         'Delete An Employee',
         'Exit']
 }];
 
+// Main function of the program
 function runEmpTracker() { 
-    inquirer.prompt(questions)
+    inquirer.prompt(question)
     .then((answer) => {
         if (answer.what === 'View All Departments') {
             viewAllDepartments();
@@ -54,6 +61,10 @@ function runEmpTracker() {
             updateEmployeeRole();
         } else if (answer.what === 'Update An Employee Manager') {
             updateEmployeeManager();
+        } else if (answer.what === 'Delete A Department') {
+            deleteDepartment();
+        } else if (answer.what === 'Delete A Role') {
+            deleteRole();
         } else if (answer.what === 'Delete An Employee') {
             deleteEmployee();
         } else if (answer.what === 'Exit') {
@@ -63,57 +74,74 @@ function runEmpTracker() {
     });
 };
 
+// Function to view Departments
 function viewAllDepartments() {
     console.log('Viewing All Departments');
     connection.query('SELECT * FROM department', (err, results) => {
+            // Shows Table
             console.table(results);
+            // Restarts inquirer after displaying the table
             runEmpTracker();
         }
       );
 };
 
+// Function to view Roles
 function viewAllRoles() {
     console.log('Viewing All Roles');
+    // Grabs just the information I want to display
     const sql = `SELECT roles.id, roles.title, roles.salary, department.dep_name AS department
 FROM roles
 JOIN department ON roles.department_id = department.id`
     connection.query(sql, (err, results) => {
+        // Shows Table
         console.table(results); 
+        // Runs it again
         runEmpTracker();
       }
     );
 };
 
+// Function to view Employees
 function viewAllEmployees() {
     console.log('Viewing All Employees');
-    const sql = `SELECT employee.first_name, employee.last_name, roles.title, roles.salary, department.dep_name, 
+    // Uses JOIN to display all the things I want, CONCAT function displays manager name instead of number
+    const sql = `SELECT employee.first_name, employee.last_name, roles.title, roles.salary, department.dep_name,
 CONCAT(e.first_name, ' ' ,e.last_name) AS Manager 
 FROM employee 
 INNER JOIN roles ON roles.id = employee.role_id 
 INNER JOIN department ON department.id = roles.department_id 
 LEFT JOIN employee e ON employee.manager_id = e.id;`
     connection.query(sql, (err, results) => {
+        // Shows Table
         console.table(results); 
+        // Runs it again
         runEmpTracker();
       }
     );
 };
 
+// Adding a new Department
 function addDepartment() {
+    // Asks for new department name
     inquirer.prompt({
         type: 'input',
         name: 'dept',
         message: 'What is the name of the new Department?'
     })
+    // Creates new department
     .then((answer) => {
         connection.query('INSERT INTO department SET ?',{dep_name: answer.dept}, (err, results) => {
+            // Shows updated Table
             viewAllDepartments();
             console.log('Successfully Added A Department');
         });
     });
 };
 
+// Adding a new Role
 function addRole() {
+    // First we fill an array with departments so the user can chose a department by name and role can be properly assigned by id 
     const departments = [];
     connection.query('SELECT * FROM department', (err, res) => {
         res.forEach(dep => {
@@ -123,6 +151,7 @@ function addRole() {
             };
             departments.push(depObj);
         });
+        // Questions abou t new Role
         inquirer.prompt([
                 {
                 type: 'input',
@@ -139,8 +168,10 @@ function addRole() {
                 choices: departments
             }
         ])
+        // Creates new Role
         .then((answer) => {
             connection.query('INSERT INTO roles SET ?',{title: answer.roleTitle, salary: answer.roleSalary, department_id: answer.roleDep}, (err, results) => {
+                // Shows updated Table
                 viewAllRoles();
                 console.log('Successfully Added A Role');
             });
@@ -148,7 +179,9 @@ function addRole() {
     });
 };
 
+// Adding an Employee
 function addEmployee() {
+    // First filling a managers array with manager objects so user can chose manager by name and the manager will be assigned by id, with there also being an option for no manager
     connection.query('SELECT * FROM employee', (err, empRes) =>{
         const managers = [
             {
@@ -162,6 +195,7 @@ function addEmployee() {
                 value: id
             });
         });
+        // Second we fill an array with departments so the user can chose a department by name and role can be properly assigned by id
         connection.query('SELECT * FROM roles', (err, roleRes) => {
             const roles = [];
             roleRes.forEach(({ title, id }) => {
@@ -170,6 +204,7 @@ function addEmployee() {
                     value: id
                 });
             });
+            // Now we ask questions about the new Employee while properly assigning IDs
             inquirer.prompt([
                 {
                     type: 'input',
@@ -191,8 +226,10 @@ function addEmployee() {
                     choices: managers
                 }
             ])
+            // Creates Employee
             .then((answers) => {
                 connection.query('INSERT INTO employee SET ?', {first_name: answers.empFirstName, last_name: answers.empLastName, role_id: answers.empRole, manager_id: answers.empManager}, (err, res) => {
+                    // Shows updated Table
                     viewAllEmployees();
                     console.log('Successfully Added An Employee');
                 });
@@ -201,8 +238,11 @@ function addEmployee() {
     });
 };
 
+// Updating an Employee Role
 function updateEmployeeRole() {
+
     connection.query('SELECT * FROM employee', (err, empRes) => {
+        // First we create employees array
         const empChoice = [];
         empRes.forEach(({ first_name, last_name, id }) => {
             empChoice.push({
@@ -210,6 +250,7 @@ function updateEmployeeRole() {
                 value: id
             });
         });
+        // Second we create a roles array
         connection.query('SELECT * FROM roles', (err, roleRes) => {
             const roles = [];
              roleRes.forEach(({ title, id }) => {
@@ -218,6 +259,7 @@ function updateEmployeeRole() {
                     value: id
                 });
              });
+            //  Using these arrays we ask which amployee will be updated to which role
              inquirer.prompt([
                 {
                     type: 'list',
@@ -231,8 +273,10 @@ function updateEmployeeRole() {
                     choices: roles
                 }
              ])
+            // Updates the Role
              .then((answers) => {
                 connection.query(`UPDATE employee SET role_id = ${answers.role} WHERE id = ${answers.employee}`, (err, res) => {
+                    // Shows updated Table
                     viewAllEmployees();
                     console.log('Successfully Updated An Employee');
                 });
@@ -241,7 +285,9 @@ function updateEmployeeRole() {
     });
 };
 
+// Updating a manager
 function updateEmployeeManager() {
+    // First we create employees array that we can use for both questions here
     connection.query('SELECT * FROM employee', (err, empRes) => {
         const empChoice = [];
         empRes.forEach(({ first_name, last_name, id }) => {
@@ -250,6 +296,7 @@ function updateEmployeeManager() {
                 value: id
             });
         });
+        // We ask which employee has a new boss
         inquirer.prompt([
             {
                 type: 'list',
@@ -263,16 +310,101 @@ function updateEmployeeManager() {
                 choices: empChoice
             }
         ])
+        // And we update the table
         .then((answers) => {
             connection.query(`UPDATE employee SET manager_id = ${answers.newManager} WHERE id = ${answers.employee}`, () => {
                 console.log('Successfully Updated The Manager For This Employee');
+                // Shows updated Table
                 viewAllEmployees();
             });
         });
     });
 };
 
+// Deletes a department, and per the schema.sql deletes roles assigned to that department, and employees in said roles
+function deleteDepartment() {
+    connection.query('SELECT * FROM department', (err, depRes) => {
+        // Departments array for the choices
+        const departments = [];
+         depRes.forEach(({ dep_name, id }) => {
+            departments.push({
+                name: dep_name,
+                value: id
+            });
+         });
+        //  Asking the Q
+         inquirer.prompt([
+            {
+                type: 'list',
+                name: 'department',
+                message: 'Which Department would you like to DELETE?',
+                choices: departments
+            },{
+                // A second question to make sure user understands what they are deleting
+                type: 'list',
+                name: 'AREYOUSURE',
+                message: 'THIS WILL DELETE THIS DEPARTMENT, ALL ROLES ASSIGNED TO THIS DEPARTMENT, AND DELETE ALL EMPLOYEES ASSIGNED TO THOSE ROLES. IF YOU ARE NOT SURE WHAT ROLES OR WHICH EMPLOYEES ARE IN THIS DEPARTMENT, SELECT "NO" AND USE "VIEW ALL EMPLOYEES" TO SEE WHO IS IN WHAT ROLE, AND THE DEPARTMENTS THEY ARE IN. IF YOU ARE SURE ABOUT DELETION, SELECT "YES,..."',
+                choices: ["NO", "YES, I  am sure I want to DELETE this DEPARTMENT, all ROLES and all EMPLOYEES in it"]
+            }
+        ])
+        // Deletes the department, roles, and employees
+         .then((answers) => {
+            if (answers.AREYOUSURE === "NO") {
+                runEmpTracker();
+            } else {
+                connection.query(`DELETE FROM department WHERE id = ${answers.department};`, (err, res) => {
+                    // Shows updated Table
+                    viewAllDepartments();
+                    console.log('Successfully Deleted A Department');
+                });
+            };
+         });
+    });
+};
+
+// Deleting a Role
+function deleteRole() {
+    // Roles array for the choices
+    connection.query('SELECT * FROM roles', (err, roleRes) => {
+        const roles = [];
+         roleRes.forEach(({ title, id }) => {
+            roles.push({
+                name: title,
+                value: id
+            });
+         });
+         inquirer.prompt([
+            {
+                type: 'list',
+                name: 'role',
+                message: 'Which role would you like to DELETE?',
+                choices: roles
+            },{
+                // The double Check
+                type: 'list',
+                name: 'AREYOUSURE',
+                message: 'THIS WILL DELETE THIS ROLE AND DELETE ALL EMPLOYEES ASSIGNED TO THIS ROLE. IF YOU ARE NOT SURE WHO IS IN THIS ROLE, SELECT "NO" AND USE "VIEW ALL EMPLOYEES" TO SEE WHO IS IN WHAT ROLE. IF YOU ARE SURE ABOUT DELETION, SELECT "YES,..."',
+                choices: ["NO", "YES, I  am sure I want to DELETE this ROLE and all EMPLOYEES in it"]
+            }
+        ])
+        // The result
+         .then((answers) => {
+            if (answers.AREYOUSURE === "NO") {
+                runEmpTracker();
+            } else {
+                connection.query(`DELETE FROM roles WHERE id = ${answers.role};`, (err, res) => {
+                    // Shows updated Table
+                    viewAllRoles();
+                    console.log('Successfully Deleted A Role');
+                });
+            };
+         });
+    });
+};
+
+// Deleting an employee
 function deleteEmployee() {
+    // Employees array
     connection.query('SELECT * FROM employee', (err, empRes) => {
         const empChoice = [];
         empRes.forEach(({ first_name, last_name, id }) => {
@@ -287,15 +419,28 @@ function deleteEmployee() {
                     name: 'employee',
                     message: 'Which Employee would you like to DELETE?',
                     choices: empChoice
+                },{
+                    // This function isn't as destructive, but still good to make sure
+                    type: 'list',
+                    name: 'AREYOUSURE',
+                    message: 'THIS WILL DELETE THIS EMPLOYEE. IF YOU ARE NOT SURE, SELECT "NO" AND USE THE PROMPTS TO DO SOMETHING ELSE. IF YOU ARE SURE ABOUT DELETION, SELECT "YES,..."',
+                    choices: ["NO", "YES, I  am sure I want to DELETE this EMPLOYEE"]
                 }
-            ])
+        ])
+        // Results are in
         .then((answers) => {
-            connection.query(`DELETE FROM employee WHERE id = ${answers.employee};`, (err, res) => {
-                viewAllEmployees();
-
-            });
+            if (answers.AREYOUSURE === "NO") {
+                runEmpTracker();
+            } else {
+                connection.query(`DELETE FROM employee WHERE id = ${answers.employee};`, (err, res) => {
+                    // Shows updated Table
+                    viewAllEmployees();
+                    console.log('Successfully Deleted An Employee');
+                });
+            };
         });
     });
 };
 
+// Runs the inquirer
 runEmpTracker();
